logfs: use uintptr_t to hold opaque pointers
When storing an opaque pointer value in a uint, the
appropriate type is uintptr_t which is guaranteed to
be wide enough to hold a pointer.

This is particularly important when the code can be
compiled in a sim or unit test environment which is
intended to run on a 64-bit machine.

Conflicts:
	flight/targets/DiscoveryF4/System/pios_board.c
	flight/targets/FlyingF4/System/pios_board.c
	flight/targets/Freedom/System/pios_board.c
	flight/targets/Quanton/System/pios_board.c

+++ flight/PiOS/Common/pios_flashfs_logfs.c
@@ -461,7 +461,7 @@ static int32_t logfs_mount_log(uint8_t arena_id)
  * @brief Initialize the flash object setting FS
  * @return 0 if success, -1 if failure
  */
-int32_t PIOS_FLASHFS_Logfs_Init(uint32_t * fs_id, const struct flashfs_logfs_cfg * cfg, const struct pios_flash_driver * driver, uint32_t flash_id)
+int32_t PIOS_FLASHFS_Logfs_Init(uintptr_t * fs_id, const struct flashfs_logfs_cfg * cfg, const struct pios_flash_driver * driver, uint32_t flash_id)
 {
 	PIOS_Assert(cfg);
 	PIOS_Assert(fs_id);
@@ -524,7 +524,7 @@ int32_t PIOS_FLASHFS_Logfs_Init(uint32_t * fs_id, const struct flashfs_logfs_cfg
 	/* Log has been mounted */
 	rc = 0;
 
-	*fs_id = (uint32_t) &logfs;
+	*fs_id = (uintptr_t) &logfs;
 
 out_end_trans:
 	logfs.driver->end_transaction(logfs.flash_id);
