+++ flight/targets/RevoMini/System/pios_board.c
@@ -303,7 +303,7 @@ void PIOS_Board_Init(void) {
 	/* Connect flash to the approrpiate interface and configure it */
 	uint32_t flash_id;
 	PIOS_Flash_Jedec_Init(&flash_id, pios_spi_telem_flash_id, 1, &flash_m25p_cfg);
-	uint32_t fs_id;
+	uintptr_t fs_id;
 	PIOS_FLASHFS_Logfs_Init(&fs_id, &flashfs_m25p_cfg, &pios_jedec_flash_driver, flash_id);
 #endif
