+++ flight/targets/Revolution/System/pios_board.c
@@ -389,7 +389,7 @@ void PIOS_Board_Init(void) {
 	uint32_t flash_id;
 	PIOS_Flash_Jedec_Init(&flash_id, pios_spi_accel_id, 1, &flash_m25p_cfg);
 #endif
-	uint32_t fs_id;
+	uintptr_t fs_id;
 	PIOS_FLASHFS_Logfs_Init(&fs_id, &flashfs_m25p_cfg, &pios_jedec_flash_driver, flash_id);
 
 	/* Initialize UAVObject libraries */
